+++ Changelog.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2022-09-29
+### Fixed
+- Changed default threshold for categorical variables.
+- Improved the description of clusters function.
+
 ## [0.1.1] - 2022-09-24
 ### Fixed
 - Fixed bug with umap.plot. Replaced by plot using matplotlib.